improve thinning, add tests
