+++ packages/dev/src/components/checkbox/checkbox.tsx
@@ -1,6 +1,10 @@
 import * as React from 'react'
 import * as Label from '@radix-ui/react-label'
-import { Root, Indicator, CheckboxOwnProps } from '@radix-ui/react-checkbox'
+import {
+  Root,
+  Indicator,
+  CheckboxProps as CheckboxOwnProps,
+} from '@radix-ui/react-checkbox'
 import styles from './checkbox.module.css'
 
 interface CheckboxProps extends CheckboxOwnProps {
