+++ packages/perfect-freehand/tsconfig.build.json
@@ -8,5 +8,13 @@
     "**/*.spec.ts",
     "src/test",
     "dist"
-  ]
+  ],
+  "compilerOptions": {
+    "composite": false,
+    "incremental": false,
+    "declarationMap": true,
+    "sourceMap": true,
+    "emitDeclarationOnly": true,
+    "stripInternal": true
+  }
 }
